Add GN build definition for Fuchsia and accomodate one test binary
This adds a GN build definition for Fuchsia and renames some duplicate
test names to allow building the tests into one binary instead of a
few dozen. See comments in BUILD.gn for details on how the folding
happens, but the renames are needed because the TEST() macro in
leveldb produces symbols based only on the test name, not the test
name + category as in gtest.

The build file hackery is pretty horrible and for the record I do feel
bad about it.

Change-Id: I971dd83f8165683d813bf7c17dffc9ce7ccd9a9b